fix(validation): When validation requests, only do type coercion on requests without implicit type
When validation a parameter in a URL, we need to do type coercion to make sure it matches the
schema, since `?foo=9` could represent the string "9" or the number 9.  We were doing type coercion
on all requests, but in a JSON request the type is explicitly specified by the JSON (e.g. `{"foo":
9}` vs. `{"foo": "9"}`).  In cases where a schem allowed either a string or a number, AJV would
sometimes do type coercion where none was neccesary.  This fixes it so that we only do type coercion
on requests with a media type of "application/x-www-form-urlencoded" or "multipart/form-data".
Since types are no longer converted in JSON requests, it's possible some clients
have been sending bad requests all this time and will now start failing.

fix #73