fix: Fix issue where we assumed json-ptrs were always in URI fragment format.
This introduces two fixes; first `jsonPaths.jsonPointerStripPrefix()`
will now try to preserve the format of a json-ptr (i.e. if you give it a
pointer in URI fragment format, it will return a pointer in URI fragment
format.)  Second, in src/oas3/Schema/validators, if we find a json-ptr
that isn't in URI fragment format, we'll treat it properly, and convert
it.

fix #132